+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0 / 2013-10-27
+
+* [FEATURE] Return new devices information from user-agent-string.info. This
+  categorizes the type of device for requests (PCs, tablets, phones, TVs, etc).
+
 ## 0.1.1 / 2013-08-02
 
 * [ENHANCEMENT] Allow only 1 process to perform the automatic data update task